+++ grapher_lib/gui_callbacks_graph_core.py
@@ -177,6 +177,7 @@ def set_dropdown_tables_for_graph(
 @callback(
     Output("memory-filtered-data", "data"),
     Output("memory-selected-tables", "data"),  # pasirinktos lentelės, bet be kaimynų
+    Output("depicted-tables-info", "children"),
     Input("tabs-container", "active_tab"),
     Input("memory-submitted-data", "data"),  # žodynas su PDSA ("node_data") ir ryšių ("edge_data") duomenimis
     Input("dropdown-tables", "value"),
@@ -207,7 +208,8 @@ def get_filtered_data_for_network(
         or active_tab != "graph"  # esame kitoje nei grafiko kortelėje
         or (not selected_dropdown_tables and not input_list_tables)  # įkelti, bet nepasirinkti
     ):
-        return {}, []
+        depicted_tables_msg = _("%d of %d") % (0, 0)
+        return {}, [], depicted_tables_msg
 
     # Visos galimos lentelės
     tables_pdsa = data_submitted["node_data"]["list_all_tables"]
@@ -300,6 +302,10 @@ def get_filtered_data_for_network(
         ]
         df_edges = pl.DataFrame(df_edges, infer_schema_length=None)
 
+    depicted_tables_msg = _("%d of %d") % (
+        len(selected_tables_and_neighbors),
+        len(tables_all) - len(tables_excludable)
+    )
 
     if df_edges.height == 0:
         df_edges = pl.DataFrame(schema={
@@ -309,7 +315,7 @@ def get_filtered_data_for_network(
         "node_elements": selected_tables_and_neighbors,
         "node_neighbors": neighbors,
         "edge_elements": df_edges.to_dicts(),  # df būtina paversti į žodyno/JSON tipą, antraip Dash nulūš
-    }, selected_tables
+    }, selected_tables, depicted_tables_msg
 
 
 @callback(